requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
requests==2.31.0